+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,11 @@ permalink: /changelog/
 * **feature-media**
   * Adds `MediaFullscreenOrientationFeature` to autorotate activity while in fullscreen based on media aspect ratio.
 
+* **support-images**
+  * ⚠️ **This is a breaking change**: Extracts `AndroidIconDecoder`, `IconDecoder` and `DesiredSize` out of `browser-icons`
+    into a new component `support-images`, which provides helpers for handling images. `AndroidIconDecoder` and `IconDecoder`
+    are renamed to `AndroidImageDecoder` and `ImageDecoder` in `support-images`.
+
 # 42.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v41.0.0...42.0.0)